Fix Channel No. 2 Zero Volume S32/Multi32
2612 Sound being used instead of ym3438, which had unfortunate side effect in Xtreme of Zero Volume for Channel No. 2.  This commit rectifies that and Channel No. 2 now has functional default sound.

Note: Any previous retroarch/saves/cfg files (for affected games) must be first deleted in order to take advantage of this fix-up!

Personal thanks Grant2258 for pinpointing the anomaly and MAME 2003 Stuff, as well as Mahoneyt944 for his tremendous efforts in the System32/Multi32/2003 MAME endeavor!  And, of course, thanks Arcadez2003, as always, for also impacting 2003 in a magnificent way:)